Align command interrupter and fix something

+++ 0-minio.sh
@@ -14,10 +14,9 @@ unzip -q awscliv2.zip
 sudo ./aws/install
 rm -rf aws awscliv2.zip
 echo "complete -C '/usr/local/bin/aws_completer' aws">/etc/profile.d/awscli.sh
+cd
 fi
 
-
-cd
 if [ ! -f /usr/local/bin/minio ]; then
 mkdir -p /minio/data
 chown minio-user:minio-user /minio
@@ -26,13 +25,19 @@ cd /minio
 wget https://dl.min.io/server/minio/release/linux-amd64/minio
 chmod +x minio
 mv minio  /usr/local/bin/
+cd
 fi
 
+if [ ! -f /usr/bin/go ]; then
 apt update
 apt -y install golang-go
+cd
+fi
+
 curl -s -o  generate_cert.go "https://golang.org/src/crypto/tls/generate_cert.go?m=text"
 IPADDR=`hostname -I | cut -d" " -f1`
 go run generate_cert.go -ca --host $IPADDR
+rm generate_cert.go
 unset IPADDR
 mv cert.pem ~/.minio/certs/public.crt
 chmod 600 ~/.minio/certs/public.crt
@@ -40,8 +45,9 @@ mv key.pem ~/.minio/certs/private.key
 chmod 600 ~/.minio/certs/private.key
 cd
 
+ if [ ! -f /etc/default/minio ]; then
 ### add minio to systemctl 
-cat <<EOT >> /etc/default/minio
+cat <<EOT > /etc/default/minio
 # Volume to be used for MinIO server.
 MINIO_VOLUMES="/minio/data"
 # Use if you want to run MinIO on a custom port.
@@ -50,11 +56,14 @@ MINIO_OPTS="--address :9000"
 MINIO_ACCESS_KEY=minioadmin
 # Secret key of the server.
 MINIO_SECRET_KEY=minioadmin
-
 EOT
+fi
+
+if [ ! -f /etc/systemd/system/minio.service ]; then
 ( cd /etc/systemd/system/; curl -O https://raw.githubusercontent.com/minio/minio-service/master/linux-systemd/minio.service )
 sed -i -e 's/minio-user/root/g' /etc/systemd/system/minio.service
 systemctl enable --now minio.service
+fi
 
 mkdir ~/.aws
 cat <<'EOF' >>~/.aws/config
@@ -77,4 +86,3 @@ echo "**************************************************************************
 echo "minio and bucket:backupkasten was created"
 echo "Next Step"
 echo "There is no action."
-
